arb-py/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "arb-py"
version = "0.1.0"
description = "Arbitration Python package"
readme = "README.md"
requires-python = ">=3.9,<4.0"
dependencies = [
    "black",
    "httpx",
    "playwright",
    "beautifulsoup4",
    "python-dotenv",
    "spidur==0.2.0",
    "openai",
]

[project.optional-dependencies]
dev = [
    "black",
]

[project.scripts]
format = "black:main"

[tool.black]
line-length = 80
target-version = ['py312']

[tool.setuptools.packages.find]
where = ["."]
include = ["arb_py*"]

[tool.poetry]
package-mode = false
